Mv quote measurement checking fn to attestation module
