+++ crates/threshold-signature-server/src/backup_provider/mod.rs
@@ -13,7 +13,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-//! Backup encryption key provider service
+//! Backup database encryption key provider service
 pub mod api;
 pub mod errors;
